// models/Company.ts
import { Coupon } from './Coupon';
import { User } from './User';


export class Company {

    compName: string;
    contactEmail: string;
    contactPhone: string;
    companyID?: number;

}





	// private long companyID;
	// private String compName;
	// private String contactEmail;
	// private String contactPhone;
	